apps/ai-code-execution/requirements.txt
streamlit>=1.32.0
e2b-code-interpreter>=0.0.8
google-generativeai>=0.5.0
anthropic>=0.25.0
openai>=1.20.0
python-dotenv>=1.0.0
pandas
openpyxl
